#96: Refactor into smaller components
