+++ src/components/Dashboard/Profile/sections/VolunteerProfileSection.tsx
@@ -1,34 +1,35 @@
 "use client";
 import Button from "@/components/core/button/Button/Button";
-import { ErrorMessage } from "@/components/core/common";
-import { Tags } from "@/components/core/common/Tags";
-import { EditableField } from "@/components/EditableField/EditableField";
-import { AvailabilityGrid } from "@/components/forms/AvailabilityGrid/AvailabilityGrid";
-import { LanguageFields } from "@/components/forms/LanguageFields";
-import { Availability } from "@/components/forms/types";
-import { getScheduleState } from "@/components/forms/utils";
 import { Heading2 } from "@/components/styled/text";
-import { apiPathOption } from "@/config/constants";
 import { useUpdateVolunteerProfile } from "@/hooks/useUpdateVolunteerProfile";
-import { LanguageLevel, LanguageObject } from "@/types";
+import { LanguageLevel } from "@/types";
 import { zodResolver } from "@hookform/resolvers/zod";
-import { UserCircle, UsersFour } from "@phosphor-icons/react";
-import { useQuery } from "@tanstack/react-query";
-import { TFunction } from "i18next";
-import {
-  Availability as ApiAvailability,
-  ApiVolunteerGet,
-  ByDay,
-  Hour,
-  Lang,
-  LangProficiency,
-  VolunteerStateTypeType,
-} from "need4deed-sdk";
-import { useCallback, useEffect, useMemo, useState } from "react";
-import { Control, Controller, ControllerRenderProps, FieldErrors, useForm, UseFormTrigger } from "react-hook-form";
+import { UserCircle } from "@phosphor-icons/react";
+import { ApiVolunteerGet, Lang, LangProficiency, VolunteerStateTypeType } from "need4deed-sdk";
+import { useEffect, useMemo, useState } from "react";
+import { useForm } from "react-hook-form";
 import { useTranslation } from "react-i18next";
 import styled from "styled-components";
-
+import { apiToFormAvailability, formToApiAvailability } from "./VolunteerProfileSection/availabilityUtils";
+import { DisplayFields } from "./VolunteerProfileSection/DisplayFields";
+import {
+  extractTitles,
+  formatActivities,
+  formatAvailability,
+  formatDistricts,
+  formatLanguages,
+  formatLanguagesForDisplay,
+  formatLocationsForDisplay,
+  formatSkills,
+  getVolunteerTypeLabel,
+} from "./VolunteerProfileSection/formatters";
+import { FormFields } from "./VolunteerProfileSection/FormFields";
+import { useApiActivities, useApiDistricts, useApiLanguages, useApiSkills } from "./VolunteerProfileSection/hooks";
+import {
+  createBiDirectionalMapping,
+  createIdToTitleMap,
+  createTitleToIdMap,
+} from "./VolunteerProfileSection/mappingUtils";
 import {
   createVolunteerProfileSchema,
   VolunteerProfileFormData,
@@ -76,53 +77,6 @@ const Details = styled.div`
   gap: 8px;
 `;
 
-const FieldRow = styled.div`
-  display: flex;
-  flex-direction: row;
-  align-items: flex-start;
-  width: 100%;
-  gap: 32px;
-  padding: 16px 0;
-  border-bottom: none;
-
-  &:last-child {
-    border-bottom: none;
-  }
-`;
-
-const FieldLabel = styled.div`
-  font-size: 20px;
-  font-weight: bold;
-  color: var(--color-midnight);
-  width: 220px;
-  flex-shrink: 0;
-`;
-
-const FieldValue = styled.div`
-  font-size: 20px;
-  color: var(--color-midnight);
-  flex: 1;
-  line-height: 1.5;
-`;
-
-const VolunteerTypeBadge = styled.div`
-  display: inline-flex;
-  align-items: center;
-  gap: 8px;
-  padding: 10px 20px;
-  background-color: var(--color-blue-500);
-  color: var(--color-white);
-  border-radius: 24px;
-  font-size: 16px;
-  font-weight: 500;
-`;
-
-const TagsWrapper = styled.div`
-  display: flex;
-  flex-wrap: wrap;
-  gap: 8px;
-`;
-
 const ButtonRow = styled.div`
   display: flex;
   flex-direction: row;
@@ -132,484 +86,29 @@ const ButtonRow = styled.div`
   width: 100%;
 `;
 
-const LanguageFieldsWrapper = styled.div`
-  margin-bottom: 8px;
-`;
-
-interface Props {
+type Props = {
   volunteer: ApiVolunteerGet;
-}
-
-// Day mapping: weekday number to ByDay enum
-const dayMap: Record<number, ByDay> = {
-  1: ByDay.MO,
-  2: ByDay.TU,
-  3: ByDay.WE,
-  4: ByDay.TH,
-  5: ByDay.FR,
-  6: ByDay.SA,
-  7: ByDay.SU,
 };
-
-const reverseDayMap: Record<string, number> = {
-  Monday: 1,
-  Tuesday: 2,
-  Wednesday: 3,
-  Thursday: 4,
-  Friday: 5,
-  Saturday: 6,
-  Sunday: 7,
-};
-
-type ApiLanguageOption = {
-  id: number;
-  title: string;
-  isoCode?: string;
-};
-
-// Hook to fetch languages from API
-function useApiLanguages(currentLanguage: Lang) {
-  return useQuery<ApiLanguageOption[]>({
-    queryKey: ["languages", currentLanguage],
-    queryFn: async () => {
-      const response = await fetch(`${apiPathOption}/language?language=${currentLanguage}`, {
-        credentials: "include",
-      });
-      if (!response.ok) throw new Error("Failed to fetch languages");
-      const data = await response.json();
-      return data.data.language || [];
-    },
-    staleTime: Infinity,
-  });
-}
-
-// Hook to fetch activities from API
-function useApiActivities(currentLanguage: Lang) {
-  return useQuery<ApiLanguageOption[]>({
-    queryKey: ["activities", currentLanguage],
-    queryFn: async () => {
-      const response = await fetch(`${apiPathOption}/activity?language=${currentLanguage}`, {
-        credentials: "include",
-      });
-      if (!response.ok) throw new Error("Failed to fetch activities");
-      const data = await response.json();
-      return data.data.activity || [];
-    },
-    staleTime: Infinity,
-  });
-}
-
-// Hook to fetch skills from API
-function useApiSkills(currentLanguage: Lang) {
-  return useQuery<ApiLanguageOption[]>({
-    queryKey: ["skills", currentLanguage],
-    queryFn: async () => {
-      const response = await fetch(`${apiPathOption}/skill?language=${currentLanguage}`, {
-        credentials: "include",
-      });
-      if (!response.ok) throw new Error("Failed to fetch skills");
-      const data = await response.json();
-      return data.data.skill || [];
-    },
-    staleTime: Infinity,
-  });
-}
-
-// Hook to fetch districts from API
-function useApiDistricts(currentLanguage: Lang) {
-  return useQuery<ApiLanguageOption[]>({
-    queryKey: ["districts", currentLanguage],
-    queryFn: async () => {
-      const response = await fetch(`${apiPathOption}/district?language=${currentLanguage}`, {
-        credentials: "include",
-      });
-      if (!response.ok) throw new Error("Failed to fetch districts");
-      const data = await response.json();
-      return data.data.district || [];
-    },
-    staleTime: Infinity,
-  });
-}
-
-// Convert API Availability[] to form Availability
-function apiToFormAvailability(apiAvailability: ApiAvailability[]): Availability {
-  const formAvailability = getScheduleState();
-
-  apiAvailability.forEach((avail) => {
-    const weekdayNum = reverseDayMap[avail.day as string];
-    if (!weekdayNum) return;
-
-    const dayIndex = formAvailability.findIndex((d) => d.weekday === weekdayNum);
-    if (dayIndex === -1) return;
-
-    // Extract time slot from daytime (e.g., ["8:00", "11:00"] -> "08-11")
-    if (Array.isArray(avail.daytime) && avail.daytime.length === 2) {
-      const [start, end] = avail.daytime;
-      const startHour = start.split(":")[0].padStart(2, "0");
-      const endHour = end.split(":")[0].padStart(2, "0");
-      const timeSlotId = `${startHour}-${endHour}`;
-
-      const slotIndex = formAvailability[dayIndex].timeSlots.findIndex((s) => s.id === timeSlotId);
-      if (slotIndex !== -1) {
-        formAvailability[dayIndex].timeSlots[slotIndex].selected = true;
-      }
-    }
-  });
-
-  return formAvailability;
-}
-
-// Convert form Availability to backend format
-function formToApiAvailability(formAvailability: Availability): Array<ApiAvailability> {
-  const result: Array<ApiAvailability> = [];
-
-  const dayEnumToString: Record<ByDay, string> = {
-    [ByDay.MO]: "Monday",
-    [ByDay.TU]: "Tuesday",
-    [ByDay.WE]: "Wednesday",
-    [ByDay.TH]: "Thursday",
-    [ByDay.FR]: "Friday",
-    [ByDay.SA]: "Saturday",
-    [ByDay.SU]: "Sunday",
-  };
-
-  formAvailability.forEach((day) => {
-    day.timeSlots.forEach((slot) => {
-      if (slot.selected && day.weekday >= 1 && day.weekday <= 7) {
-        const dayEnum = dayMap[day.weekday];
-        const dayString = dayEnumToString[dayEnum];
-        const slotId = String(slot.id);
-        const [startHourStr, endHourStr] = slotId.split("-");
-        const startHourNum = parseInt(startHourStr, 10);
-        const endHourNum = parseInt(endHourStr, 10);
-
-        const startTime = `${startHourNum}:00`;
-        const endTime = `${endHourNum}:00`;
-
-        result.push({
-          day: dayString as ByDay,
-          daytime: [startTime as Hour, endTime as Hour],
-          timeslotId: 0,
-          id: `${dayString}-${slotId}`,
-          description: "Volunteer Availability",
-          start: new Date(),
-        });
-      }
-    });
-  });
-
-  return result;
-}
-
-type DisplayFieldsProps = {
-  languages: string;
-  availability: string;
-  districts: string;
-  volunteerType: string;
-  activities: string[];
-  skills: string[];
-  t: (key: string) => string;
-};
-
-function DisplayFields({
-  languages,
-  availability,
-  districts,
-  volunteerType,
-  activities,
-  skills,
-  t,
-}: DisplayFieldsProps) {
-  return (
-    <>
-      <FieldRow>
-        <FieldLabel>{t("dashboard.volunteerProfile.profileSection.languages")}</FieldLabel>
-        <FieldValue>{languages}</FieldValue>
-      </FieldRow>
-
-      <FieldRow>
-        <FieldLabel>{t("dashboard.volunteerProfile.profileSection.availability")}</FieldLabel>
-        <FieldValue>{availability}</FieldValue>
-      </FieldRow>
-
-      <FieldRow>
-        <FieldLabel>{t("dashboard.volunteerProfile.profileSection.districts")}</FieldLabel>
-        <FieldValue>{districts}</FieldValue>
-      </FieldRow>
-
-      <FieldRow>
-        <FieldLabel>{t("dashboard.volunteerProfile.profileSection.volunteerType")}</FieldLabel>
-        <FieldValue>
-          {volunteerType ? (
-            <VolunteerTypeBadge>
-              <UsersFour size={20} weight="fill" />
-              {volunteerType}
-            </VolunteerTypeBadge>
-          ) : (
-            "–"
-          )}
-        </FieldValue>
-      </FieldRow>
-
-      <FieldRow>
-        <FieldLabel>{t("dashboard.volunteerProfile.profileSection.activities")}</FieldLabel>
-        <FieldValue>
-          {activities.length > 0 ? (
-            <TagsWrapper>
-              <Tags tags={activities} backgroundColor="var(--color-pink-100)" />
-            </TagsWrapper>
-          ) : (
-            "–"
-          )}
-        </FieldValue>
-      </FieldRow>
-
-      <FieldRow>
-        <FieldLabel>{t("dashboard.volunteerProfile.profileSection.skills")}</FieldLabel>
-        <FieldValue>
-          {skills.length > 0 ? (
-            <TagsWrapper>
-              <Tags tags={skills} backgroundColor="var(--color-pink-50)" />
-            </TagsWrapper>
-          ) : (
-            "–"
-          )}
-        </FieldValue>
-      </FieldRow>
-    </>
-  );
-}
-
-type FormFieldsProps = {
-  control: Control<VolunteerProfileFormData>;
-  errors: FieldErrors<VolunteerProfileFormData>;
-  t: TFunction<"translation", undefined>;
-  i18n: { language: string };
-  locationOptions: string[];
-  idToLabel: Record<string | number, string>;
-  labelToId: Record<string, string | number>;
-  activitiesOptions: string[];
-  activityIdToLabel: Record<string | number, string>;
-  activityLabelToId: Record<string, string | number>;
-  skillsOptions: string[];
-  skillIdToLabel: Record<string | number, string>;
-  skillLabelToId: Record<string, string | number>;
-  languagesForForm: Array<{ id: number | string; title: Record<Lang, string> }>;
-  trigger: UseFormTrigger<VolunteerProfileFormData>;
-};
-
-function FormFields({
-  control,
-  errors,
-  t,
-  i18n,
-  locationOptions,
-  idToLabel,
-  labelToId,
-  activitiesOptions,
-  activityIdToLabel,
-  activityLabelToId,
-  skillsOptions,
-  skillIdToLabel,
-  skillLabelToId,
-  languagesForForm,
-  trigger,
-}: FormFieldsProps) {
-  return (
-    <>
-      <Controller
-        name="languages"
-        control={control}
-        render={({ field }: { field: ControllerRenderProps<VolunteerProfileFormData, "languages"> }) => (
-          <FieldRow>
-            <FieldLabel>{t("dashboard.volunteerProfile.profileSection.languages")}:</FieldLabel>
-            <FieldValue>
-              <LanguageFieldsWrapper>
-                <LanguageFields
-                  languages={field.value}
-                  onChange={field.onChange}
-                  t={t}
-                  availableLanguages={languagesForForm}
-                />
-              </LanguageFieldsWrapper>
-              {errors.languages?.message && <ErrorMessage message={errors.languages.message} />}
-            </FieldValue>
-          </FieldRow>
-        )}
-      />
-
-      <Controller
-        name="availability"
-        control={control}
-        render={({ field }: { field: ControllerRenderProps<VolunteerProfileFormData, "availability"> }) => (
-          <FieldRow>
-            <FieldLabel>{t("dashboard.volunteerProfile.profileSection.availability")}:</FieldLabel>
-            <FieldValue>
-              <AvailabilityGrid
-                availability={field.value}
-                onChange={field.onChange}
-                t={t}
-                currentLanguage={i18n.language as Lang}
-              />
-              {errors.availability?.message && <ErrorMessage message={errors.availability.message} />}
-            </FieldValue>
-          </FieldRow>
-        )}
-      />
-
-      <Controller
-        name="districts"
-        control={control}
-        render={({ field }: { field: ControllerRenderProps<VolunteerProfileFormData, "districts"> }) => (
-          <EditableField
-            mode="edit"
-            type="checkbox-list"
-            label={t("dashboard.volunteerProfile.profileSection.districts")}
-            value={field.value.map((id) => idToLabel[id] || String(id))}
-            setValue={(value) => {
-              const labels = Array.isArray(value) ? value : [value];
-              field.onChange(labels.map((label) => String(labelToId[label])));
-              trigger("districts");
-            }}
-            options={locationOptions}
-            errorMessage={errors.districts?.message}
-          />
-        )}
-      />
-
-      <Controller
-        name="volunteerType"
-        control={control}
-        render={({ field }: { field: ControllerRenderProps<VolunteerProfileFormData, "volunteerType"> }) => (
-          <EditableField
-            mode="edit"
-            type="radio-list"
-            label={t("dashboard.volunteerProfile.profileSection.volunteerType")}
-            value={field.value}
-            setValue={field.onChange}
-            options={Object.values(VolunteerStateTypeType)
-              .filter((type): type is VolunteerStateTypeType => type !== undefined)
-              .map((type) => t(`dashboard.volunteerProfile.volunteerHeader.volunteerType_options.${type}`))}
-            errorMessage={errors.volunteerType?.message}
-          />
-        )}
-      />
-
-      <Controller
-        name="activities"
-        control={control}
-        render={({ field }: { field: ControllerRenderProps<VolunteerProfileFormData, "activities"> }) => (
-          <EditableField
-            mode="edit"
-            type="checkbox-list"
-            label={t("dashboard.volunteerProfile.profileSection.activities")}
-            value={field.value.map((id) => activityIdToLabel[id] || String(id))}
-            setValue={(value) => {
-              const labels = Array.isArray(value) ? value : [value];
-              field.onChange(labels.map((label) => String(activityLabelToId[label])));
-              trigger("activities");
-            }}
-            options={activitiesOptions}
-            errorMessage={errors.activities?.message}
-          />
-        )}
-      />
-
-      <Controller
-        name="skills"
-        control={control}
-        render={({ field }: { field: ControllerRenderProps<VolunteerProfileFormData, "skills"> }) => (
-          <EditableField
-            mode="edit"
-            type="checkbox-list"
-            label={t("dashboard.volunteerProfile.profileSection.skills")}
-            value={field.value.map((id) => skillIdToLabel[id] || String(id))}
-            setValue={(value) => {
-              const labels = Array.isArray(value) ? value : [value];
-              field.onChange(labels.map((label) => String(skillLabelToId[label])));
-              trigger("skills");
-            }}
-            options={skillsOptions}
-            errorMessage={errors.skills?.message}
-          />
-        )}
-      />
-    </>
-  );
-}
 
 export function VolunteerProfileSection({ volunteer }: Props) {
   const { t, i18n } = useTranslation();
   const { mutate: updateProfile, isPending } = useUpdateVolunteerProfile(volunteer.id);
   const [isEditing, setIsEditing] = useState(false);
 
-  // Fetch options from API
   const { data: apiLanguages = [] } = useApiLanguages(i18n.language as Lang);
   const { data: apiActivities = [] } = useApiActivities(i18n.language as Lang);
   const { data: apiSkills = [] } = useApiSkills(i18n.language as Lang);
   const { data: apiDistricts = [] } = useApiDistricts(i18n.language as Lang);
 
-  // Create a lookup map from volunteer's language titles to database IDs
-  const languageTitleToDbId = useMemo(() => {
-    const map: Record<string, number> = {};
-    apiLanguages.forEach((lang) => {
-      map[lang.title.toLowerCase()] = lang.id;
-    });
-    return map;
-  }, [apiLanguages]);
+  const languageTitleToDbId = useMemo(() => createTitleToIdMap(apiLanguages), [apiLanguages]);
+  const languageIdToTitle = useMemo(() => createIdToTitleMap(apiLanguages), [apiLanguages]);
+  const activityTitleToDbId = useMemo(() => createTitleToIdMap(apiActivities), [apiActivities]);
+  const activityIdToTitle = useMemo(() => createIdToTitleMap(apiActivities), [apiActivities]);
+  const skillTitleToDbId = useMemo(() => createTitleToIdMap(apiSkills), [apiSkills]);
+  const skillIdToTitle = useMemo(() => createIdToTitleMap(apiSkills), [apiSkills]);
+  const districtTitleToDbId = useMemo(() => createTitleToIdMap(apiDistricts), [apiDistricts]);
+  const districtIdToTitle = useMemo(() => createIdToTitleMap(apiDistricts), [apiDistricts]);
 
-  // Create activity title to ID and ID to title mappings
-  const activityTitleToDbId = useMemo(() => {
-    const map: Record<string, number> = {};
-    apiActivities.forEach((act) => {
-      map[act.title.toLowerCase()] = act.id;
-    });
-    return map;
-  }, [apiActivities]);
-
-  const activityIdToTitle = useMemo(() => {
-    const map: Record<number, string> = {};
-    apiActivities.forEach((act) => {
-      map[act.id] = act.title;
-    });
-    return map;
-  }, [apiActivities]);
-
-  // Create skill title to ID and ID to title mappings
-  const skillTitleToDbId = useMemo(() => {
-    const map: Record<string, number> = {};
-    apiSkills.forEach((skill) => {
-      map[skill.title.toLowerCase()] = skill.id;
-    });
-    return map;
-  }, [apiSkills]);
-
-  const skillIdToTitle = useMemo(() => {
-    const map: Record<number, string> = {};
-    apiSkills.forEach((skill) => {
-      map[skill.id] = skill.title;
-    });
-    return map;
-  }, [apiSkills]);
-
-  // Create district title to ID and ID to title mappings
-  const districtTitleToDbId = useMemo(() => {
-    const map: Record<string, number> = {};
-    apiDistricts.forEach((district) => {
-      map[district.title.toLowerCase()] = district.id;
-    });
-    return map;
-  }, [apiDistricts]);
-
-  const districtIdToTitle = useMemo(() => {
-    const map: Record<number, string> = {};
-    apiDistricts.forEach((district) => {
-      map[district.id] = district.title;
-    });
-    return map;
-  }, [apiDistricts]);
-
-  // Convert API languages to the format expected by LanguageFieldRow
   const languagesForForm = useMemo(() => {
     return apiLanguages.map((lang) => ({
       id: lang.id,
@@ -617,215 +116,32 @@ export function VolunteerProfileSection({ volunteer }: Props) {
     }));
   }, [apiLanguages, i18n.language]);
 
-  const locationOptions = apiDistricts.map((district) => district.title);
+  const locationOptions = useMemo(() => apiDistricts.map((district) => district.title), [apiDistricts]);
+  const activitiesOptions = useMemo(() => apiActivities.map((activity) => activity.title), [apiActivities]);
+  const skillsOptions = useMemo(() => apiSkills.map((skill) => skill.title), [apiSkills]);
 
-  const idToLabel: Record<string | number, string> = {};
-  const labelToId: Record<string, string | number> = {};
-  apiDistricts.forEach((district) => {
-    idToLabel[district.id] = district.title;
-    labelToId[district.title] = district.id;
-  });
+  const districtMapping = useMemo(() => createBiDirectionalMapping(apiDistricts), [apiDistricts]);
+  const activityMapping = useMemo(() => createBiDirectionalMapping(apiActivities), [apiActivities]);
+  const skillMapping = useMemo(() => createBiDirectionalMapping(apiSkills), [apiSkills]);
 
-  const activitiesOptions = apiActivities.map((activity) => activity.title);
-
-  const activityIdToLabel: Record<string | number, string> = {};
-  const activityLabelToId: Record<string, string | number> = {};
-  apiActivities.forEach((activity) => {
-    activityIdToLabel[activity.id] = activity.title;
-    activityLabelToId[activity.title] = activity.id;
-  });
-
-  const skillsOptions = apiSkills.map((skill) => skill.title);
-
-  const skillIdToLabel: Record<string | number, string> = {};
-  const skillLabelToId: Record<string, string | number> = {};
-  apiSkills.forEach((skill) => {
-    skillIdToLabel[skill.id] = skill.title;
-    skillLabelToId[skill.title] = skill.id;
-  });
-
-  // Create ID to title mapping for display
-  const languageIdToTitle = useMemo(() => {
-    const map: Record<number, string> = {};
-    apiLanguages.forEach((lang) => {
-      map[lang.id] = lang.title;
-    });
-    return map;
-  }, [apiLanguages]);
-
-  // Transform API languages to form format
-  const formatLanguages = useCallback(
-    (langs: typeof volunteer.languages): LanguageObject[] => {
-      if (!langs || langs.length === 0) {
-        return [{ id: 1, language: "", level: "" }];
-      }
-
-      const proficiencyToLevel: Record<string, LanguageLevel> = {
-        native: LanguageLevel.NATIVE,
-        fluent: LanguageLevel.FLUENT,
-        intermediate: LanguageLevel.INTERMEDIATE,
-        advanced: LanguageLevel.FLUENT,
-        beginner: LanguageLevel.INTERMEDIATE,
-      };
-
-      return langs.map((lang) => {
-        // Find the database ID by matching the title (case-insensitive)
-        const dbId = languageTitleToDbId[lang.title.toLowerCase()] || lang.id;
-
-        return {
-          id: lang.id,
-          language: String(dbId),
-          level: proficiencyToLevel[lang.proficiency?.toLowerCase() || "native"] || LanguageLevel.NATIVE,
-        };
-      });
-    },
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-    [languageTitleToDbId],
-  );
-
-  // Format activities for form - convert from API response to form IDs
-  const formatActivities = useCallback(
-    (activities: typeof volunteer.activities): string[] => {
-      if (!activities || activities.length === 0) return [];
-      return activities
-        .map((act) => {
-          // Try to match by title first, fallback to direct ID
-          const dbId = activityTitleToDbId[act.title.toLowerCase()] || act.id;
-          return String(dbId);
-        })
-        .filter((id) => id && id !== "0");
-    },
-    [activityTitleToDbId, volunteer],
-  );
-
-  // Format skills for form - convert from API response to form IDs
-  const formatSkills = useCallback(
-    (skills: typeof volunteer.skills): string[] => {
-      if (!skills || skills.length === 0) return [];
-      return skills
-        .map((skill) => {
-          // Try to match by title first, fallback to direct ID
-          const dbId = skillTitleToDbId[skill.title.toLowerCase()] || skill.id;
-          return String(dbId);
-        })
-        .filter((id) => id && id !== "0");
-    },
-    [skillTitleToDbId, volunteer],
-  );
-
-  // Format districts for form - convert from API response to form IDs
-  const formatDistricts = useCallback(
-    (locations: typeof volunteer.locations): string[] => {
-      if (!locations || locations.length === 0) return [];
-      return locations
-        .map((loc) => {
-          // Try to match by title first, fallback to direct ID
-          const dbId = districtTitleToDbId[loc.title.toLowerCase()] || loc.id;
-          return String(dbId);
-        })
-        .filter((id) => id && id !== "0");
-    },
-    [districtTitleToDbId, volunteer],
-  );
-
-  // Format languages for display
-  const formatLanguagesForDisplay = useCallback(
-    (langs: typeof volunteer.languages): string => {
-      if (!langs || langs.length === 0) return "–";
-      return langs
-        .map((lang) => {
-          // Use localized language title from API
-          const localizedTitle = languageIdToTitle[lang.id] || lang.title;
-          // Use localized proficiency level from translations
-          const proficiencyKey = lang.proficiency?.toLowerCase();
-          const localizedProficiency = proficiencyKey
-            ? t(`dashboard.volunteers.langProficiency.${proficiencyKey}`)
-            : "";
-          const proficiency = localizedProficiency ? ` – ${localizedProficiency}` : "";
-          return `${localizedTitle}${proficiency}`;
-        })
-        .join(", ");
-    },
-    [languageIdToTitle, t, volunteer],
-  );
-
-  const formatAvailability = useCallback(
-    (avails: typeof volunteer.availability): string => {
-      if (!avails || avails.length === 0) {
-        const defaultSchedule = getScheduleState();
-        const hasSelectedSlots = defaultSchedule.some((day) => day.timeSlots.some((slot) => slot.selected));
-        if (!hasSelectedSlots) return "–";
-      }
-
-      // Group by time slots
-      const timeSlotGroups = new Map<string, string[]>();
-
-      avails.forEach((avail) => {
-        const dayName = avail.day === "occasionally" ? "Occasionally" : avail.day;
-        const timeKey =
-          Array.isArray(avail.daytime) && avail.daytime.length === 2
-            ? `${avail.daytime[0]}-${avail.daytime[1]}`
-            : avail.daytime[0] || "";
-
-        if (!timeSlotGroups.has(timeKey)) {
-          timeSlotGroups.set(timeKey, []);
-        }
-        timeSlotGroups.get(timeKey)?.push(dayName);
-      });
-
-      // Format each time slot group
-      const formatted = Array.from(timeSlotGroups.entries()).map(([time, days]) => {
-        const daysStr = days.join(" & ");
-        return time ? `${daysStr}, ${time}` : daysStr;
-      });
-
-      return formatted.join("; ");
-    },
-    [volunteer],
-  );
-
-  const extractTitles = useCallback((items: { title: string }[]): string[] => {
-    if (!items || items.length === 0) return [];
-    return items.map((item) => item.title);
-  }, []);
-
-  const getVolunteerTypeLabel = useCallback(
-    (statusType: VolunteerStateTypeType | undefined): string => {
-      if (!statusType) return "";
-
-      return t(`dashboard.volunteerProfile.volunteerHeader.volunteerType_options.${statusType}`);
-    },
-    [t],
-  );
-
-  const formatLocationsForDisplay = useCallback(
-    (locations: typeof volunteer.locations): string => {
-      if (!locations || locations.length === 0) return "–";
-      return locations.map((loc) => loc.title).join(", ");
-    },
-    [volunteer],
-  );
-
-  const schema = createVolunteerProfileSchema(t);
+  const schema = useMemo(() => createVolunteerProfileSchema(t), [t]);
 
   const { control, handleSubmit, reset, trigger, formState } = useForm<VolunteerProfileFormData>({
     resolver: zodResolver(schema),
     mode: "onChange",
     defaultValues: {
-      languages: formatLanguages(volunteer.languages),
+      languages: formatLanguages(volunteer.languages, languageTitleToDbId),
       availability: apiToFormAvailability(volunteer.availability),
-      districts: formatDistricts(volunteer.locations),
-      volunteerType: getVolunteerTypeLabel(volunteer.statusType),
-      activities: formatActivities(volunteer.activities),
-      skills: formatSkills(volunteer.skills),
+      districts: formatDistricts(volunteer.locations, districtTitleToDbId),
+      volunteerType: getVolunteerTypeLabel(volunteer.statusType, t),
+      activities: formatActivities(volunteer.activities, activityTitleToDbId),
+      skills: formatSkills(volunteer.skills, skillTitleToDbId),
     },
   });
 
   const { errors, isValid, isDirty } = formState;
 
-  // Sync form when volunteer data changes (after refetch) OR when API data loads
   useEffect(() => {
-    // Wait for API data to load before resetting form
     if (
       apiLanguages.length === 0 ||
       apiActivities.length === 0 ||
@@ -836,25 +152,24 @@ export function VolunteerProfileSection({ volunteer }: Props) {
     }
 
     reset({
-      languages: formatLanguages(volunteer.languages),
+      languages: formatLanguages(volunteer.languages, languageTitleToDbId),
       availability: apiToFormAvailability(volunteer.availability),
-      districts: formatDistricts(volunteer.locations),
-      volunteerType: getVolunteerTypeLabel(volunteer.statusType),
-      activities: formatActivities(volunteer.activities),
-      skills: formatSkills(volunteer.skills),
+      districts: formatDistricts(volunteer.locations, districtTitleToDbId),
+      volunteerType: getVolunteerTypeLabel(volunteer.statusType, t),
+      activities: formatActivities(volunteer.activities, activityTitleToDbId),
+      skills: formatSkills(volunteer.skills, skillTitleToDbId),
     });
-    // Trigger validation after reset to update isValid state
     trigger();
     setIsEditing(false);
   }, [
     volunteer,
     reset,
     trigger,
-    formatLanguages,
-    formatActivities,
-    formatSkills,
-    formatDistricts,
-    getVolunteerTypeLabel,
+    t,
+    languageTitleToDbId,
+    activityTitleToDbId,
+    skillTitleToDbId,
+    districtTitleToDbId,
     apiLanguages,
     apiActivities,
     apiSkills,
@@ -877,67 +192,45 @@ export function VolunteerProfileSection({ volunteer }: Props) {
       [LanguageLevel.INTERMEDIATE]: LangProficiency.INTERMEDIATE,
     };
 
-    const labelToVolunteerType = Object.values(VolunteerStateTypeType).reduce((acc, type) => {
-      acc[t(`dashboard.volunteerProfile.volunteerHeader.volunteerType_options.${type}`)] = type;
-      return acc;
-    }, {} as Record<string, VolunteerStateTypeType>);
+    const labelToVolunteerType = Object.values(VolunteerStateTypeType).reduce(
+      (acc, type) => {
+        acc[t(`dashboard.volunteerProfile.volunteerHeader.volunteerType_options.${type}`)] = type;
+        return acc;
+      },
+      {} as Record<string, VolunteerStateTypeType>,
+    );
 
     const statusType = labelToVolunteerType[data.volunteerType];
-
-    const validStatusTypes = Object.values(VolunteerStateTypeType);
-    const isValidStatusType = statusType && validStatusTypes.includes(statusType);
+    const isValidStatusType = statusType && Object.values(VolunteerStateTypeType).includes(statusType);
 
     updateProfile(
       {
-        // Cast to any - SDK types expect Hour enums but backend actually expects time strings
         availability: formToApiAvailability(data.availability),
         ...(isValidStatusType ? { statusType } : {}),
         languages: data.languages
-          .filter(
-            (lang): lang is LanguageObject & { level: LanguageLevel } => lang.language !== "" && lang.level !== "",
-          )
-          .map((lang) => {
-            const dbId = parseInt(lang.language, 10);
-            const languageTitle = languageIdToTitle[dbId] || "";
-
-            return {
-              id: dbId,
-              title: languageTitle,
-              proficiency: levelToProficiency[lang.level],
-            };
-          }),
+          .filter((lang) => lang.language !== "" && lang.level !== "")
+          .map((lang) => ({
+            id: parseInt(lang.language, 10),
+            title: languageIdToTitle[parseInt(lang.language, 10)] || "",
+            proficiency: levelToProficiency[lang.level as LanguageLevel],
+          })),
         locations: data.districts
-          .map((districtId) => {
-            const dbId = parseInt(districtId, 10);
-            const districtTitle = districtIdToTitle[dbId] || "";
-
-            return {
-              id: dbId,
-              title: districtTitle,
-            };
-          })
+          .map((districtId) => ({
+            id: parseInt(districtId, 10),
+            title: districtIdToTitle[parseInt(districtId, 10)] || "",
+          }))
           .filter((loc) => !isNaN(loc.id) && loc.id > 0),
         activities: data.activities
-          .map((activityId) => {
-            const dbId = parseInt(activityId, 10);
-            const activityTitle = activityIdToTitle[dbId] || "";
-
-            return {
-              id: dbId,
-              title: activityTitle,
-            };
-          })
+          .map((activityId) => ({
+            id: parseInt(activityId, 10),
+            title: activityIdToTitle[parseInt(activityId, 10)] || "",
+          }))
           .filter((act) => !isNaN(act.id) && act.id > 0),
         skills: data.skills
-          .map((skillId) => {
-            const dbId = parseInt(skillId, 10);
-            const skillTitle = skillIdToTitle[dbId] || "";
-
-            return {
-              id: dbId,
-              title: skillTitle,
-            };
-          })
+          .map((skillId) => ({
+            id: parseInt(skillId, 10),
+            title: skillIdToTitle[parseInt(skillId, 10)] || "",
+          }))
           .filter((s) => !isNaN(s.id) && s.id > 0),
       },
       {
@@ -979,23 +272,23 @@ export function VolunteerProfileSection({ volunteer }: Props) {
             t={t}
             i18n={i18n}
             locationOptions={locationOptions}
-            idToLabel={idToLabel}
-            labelToId={labelToId}
+            idToLabel={districtMapping.idToLabel}
+            labelToId={districtMapping.labelToId}
             activitiesOptions={activitiesOptions}
-            activityIdToLabel={activityIdToLabel}
-            activityLabelToId={activityLabelToId}
+            activityIdToLabel={activityMapping.idToLabel}
+            activityLabelToId={activityMapping.labelToId}
             skillsOptions={skillsOptions}
-            skillIdToLabel={skillIdToLabel}
-            skillLabelToId={skillLabelToId}
+            skillIdToLabel={skillMapping.idToLabel}
+            skillLabelToId={skillMapping.labelToId}
             languagesForForm={languagesForForm}
             trigger={trigger}
           />
         ) : (
           <DisplayFields
-            languages={formatLanguagesForDisplay(volunteer.languages)}
+            languages={formatLanguagesForDisplay(volunteer.languages, languageIdToTitle, t)}
             availability={formatAvailability(volunteer.availability)}
             districts={formatLocationsForDisplay(volunteer.locations)}
-            volunteerType={getVolunteerTypeLabel(volunteer.statusType)}
+            volunteerType={getVolunteerTypeLabel(volunteer.statusType, t)}
             activities={extractTitles(volunteer.activities)}
             skills={extractTitles(volunteer.skills)}
             t={t}
